refactor shared api types to common package
