+++ .golangci.yml
@@ -219,6 +219,13 @@ linters:
           - staticcheck
         text: 'SA1019: .+ is deprecated: Use Composition Functions instead.'
 
+      # Some shared structs in apis/common/v1 are moved to
+      # apis/common. To preserve a backward-compatible directory structure
+      # package had to be named common, which we suppress.
+      - linters:
+          - revive
+        text: "var-naming: avoid meaningless package names"
+        path: apis/common
     paths:
       - zz_generated\..+\.go$
       - .+\.pb.go$
