+++ apis/common/doc.go
@@ -1,0 +1,3 @@
+// Package common contains core API types used by most Crossplane resources.
+// +kubebuilder:object:generate=true
+package common
