+++ apis/common/v1/resource.go
@@ -18,8 +18,8 @@ package v1
 
 import (
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/types"
+
+	"github.com/crossplane/crossplane-runtime/apis/common"
 )
 
 const (
@@ -45,11 +45,11 @@ const (
 
 // LabelKeyProviderKind is added to ProviderConfigUsages to relate them to their
 // ProviderConfig.
-const LabelKeyProviderKind = "crossplane.io/provider-config-kind"
+const LabelKeyProviderKind = common.LabelKeyProviderKind
 
 // LabelKeyProviderName is added to ProviderConfigUsages to relate them to their
 // ProviderConfig.
-const LabelKeyProviderName = "crossplane.io/provider-config"
+const LabelKeyProviderName = common.LabelKeyProviderName
 
 // NOTE(negz): The below secret references differ from ObjectReference and
 // LocalObjectReference in that they include only the fields Crossplane needs to
@@ -58,180 +58,41 @@ const LabelKeyProviderName = "crossplane.io/provider-config"
 
 // A LocalSecretReference is a reference to a secret in the same namespace as
 // the referencer.
-type LocalSecretReference struct {
-	// Name of the secret.
-	Name string `json:"name"`
-}
+type LocalSecretReference = common.LocalSecretReference
 
 // A SecretReference is a reference to a secret in an arbitrary namespace.
-type SecretReference struct {
-	// Name of the secret.
-	Name string `json:"name"`
-	// Namespace of the secret.
-	Namespace string `json:"namespace"`
-}
+type SecretReference = common.SecretReference
 
 // A SecretKeySelector is a reference to a secret key in an arbitrary namespace.
-type SecretKeySelector struct {
-	SecretReference `json:",inline"`
-
-	// The key to select.
-	Key string `json:"key"`
-}
+type SecretKeySelector = common.SecretKeySelector
 
 // A LocalSecretKeySelector is a reference to a secret key
 // in the same namespace with the referencing object.
-type LocalSecretKeySelector struct {
-	LocalSecretReference `json:",inline"`
-
-	Key string `json:"key"`
-}
-
-// ToSecretKeySelector is a convenience method for converting the
-// LocalSecretKeySelector to a SecretKeySelector with the given namespace.
-func (ls *LocalSecretKeySelector) ToSecretKeySelector(namespace string) *SecretKeySelector {
-	return &SecretKeySelector{
-		SecretReference: SecretReference{
-			Name:      ls.Name,
-			Namespace: namespace,
-		},
-		Key: ls.Key,
-	}
-}
+type LocalSecretKeySelector = common.LocalSecretKeySelector
 
 // Policy represents the Resolve and Resolution policies of Reference instance.
-type Policy struct {
-	// Resolve specifies when this reference should be resolved. The default
-	// is 'IfNotPresent', which will attempt to resolve the reference only when
-	// the corresponding field is not present. Use 'Always' to resolve the
-	// reference on every reconcile.
-	// +optional
-	// +kubebuilder:validation:Enum=Always;IfNotPresent
-	Resolve *ResolvePolicy `json:"resolve,omitempty"`
-
-	// Resolution specifies whether resolution of this reference is required.
-	// The default is 'Required', which means the reconcile will fail if the
-	// reference cannot be resolved. 'Optional' means this reference will be
-	// a no-op if it cannot be resolved.
-	// +optional
-	// +kubebuilder:default=Required
-	// +kubebuilder:validation:Enum=Required;Optional
-	Resolution *ResolutionPolicy `json:"resolution,omitempty"`
-}
-
-// IsResolutionPolicyOptional checks whether the resolution policy of relevant reference is Optional.
-func (p *Policy) IsResolutionPolicyOptional() bool {
-	if p == nil || p.Resolution == nil {
-		return false
-	}
-
-	return *p.Resolution == ResolutionPolicyOptional
-}
-
-// IsResolvePolicyAlways checks whether the resolution policy of relevant reference is Always.
-func (p *Policy) IsResolvePolicyAlways() bool {
-	if p == nil || p.Resolve == nil {
-		return false
-	}
-
-	return *p.Resolve == ResolvePolicyAlways
-}
+type Policy = common.Policy
 
 // A Reference to a named object.
-type Reference struct {
-	// Name of the referenced object.
-	Name string `json:"name"`
-
-	// Policies for referencing.
-	// +optional
-	Policy *Policy `json:"policy,omitempty"`
-}
+type Reference = common.Reference
 
 // A NamespacedReference to a named object.
-type NamespacedReference struct {
-	// Name of the referenced object.
-	Name string `json:"name"`
-
-	// Namespace of the referenced object
-	// +optional
-	Namespace string `json:"namespace,omitempty"`
-	// Policies for referencing.
-	// +optional
-	Policy *Policy `json:"policy,omitempty"`
-}
+type NamespacedReference = common.NamespacedReference
 
 // A TypedReference refers to an object by Name, Kind, and APIVersion. It is
 // commonly used to reference cluster-scoped objects or objects where the
 // namespace is already known.
-type TypedReference struct {
-	// APIVersion of the referenced object.
-	APIVersion string `json:"apiVersion"`
-
-	// Kind of the referenced object.
-	Kind string `json:"kind"`
-
-	// Name of the referenced object.
-	Name string `json:"name"`
-
-	// UID of the referenced object.
-	// +optional
-	UID types.UID `json:"uid,omitempty"`
-}
+type TypedReference = common.TypedReference
 
 // A Selector selects an object.
-type Selector struct {
-	// MatchLabels ensures an object with matching labels is selected.
-	MatchLabels map[string]string `json:"matchLabels,omitempty"`
-
-	// MatchControllerRef ensures an object with the same controller reference
-	// as the selecting object is selected.
-	MatchControllerRef *bool `json:"matchControllerRef,omitempty"`
-
-	// Policies for selection.
-	// +optional
-	Policy *Policy `json:"policy,omitempty"`
-}
+type Selector = common.Selector
 
 // NamespacedSelector selects a namespaced object.
-type NamespacedSelector struct {
-	// MatchLabels ensures an object with matching labels is selected.
-	MatchLabels map[string]string `json:"matchLabels,omitempty"`
-
-	// MatchControllerRef ensures an object with the same controller reference
-	// as the selecting object is selected.
-	MatchControllerRef *bool `json:"matchControllerRef,omitempty"`
-
-	// Policies for selection.
-	// +optional
-	Policy *Policy `json:"policy,omitempty"`
-
-	// Namespace for the selector
-	// +optional
-	Namespace string `json:"namespace,omitempty"`
-}
+type NamespacedSelector = common.NamespacedSelector
 
 // ProviderConfigReference is a typed reference to a ProviderConfig
 // object, with a known api group.
-type ProviderConfigReference struct {
-	// Kind of the referenced object.
-	Kind string `json:"kind"`
-
-	// Name of the referenced object.
-	Name string `json:"name"`
-}
-
-// SetGroupVersionKind sets the Kind and APIVersion of a TypedReference.
-func (obj *TypedReference) SetGroupVersionKind(gvk schema.GroupVersionKind) {
-	obj.APIVersion, obj.Kind = gvk.ToAPIVersionAndKind()
-}
-
-// GroupVersionKind gets the GroupVersionKind of a TypedReference.
-func (obj *TypedReference) GroupVersionKind() schema.GroupVersionKind {
-	return schema.FromAPIVersionAndKind(obj.APIVersion, obj.Kind)
-}
-
-// GetObjectKind get the ObjectKind of a TypedReference.
-func (obj *TypedReference) GetObjectKind() schema.ObjectKind { return obj }
+type ProviderConfigReference = common.ProviderConfigReference
 
 // TODO(negz): Rename Resource* to Managed* to clarify that they enable the
 // resource.Managed interface.
@@ -285,80 +146,48 @@ type ResourceStatus struct {
 
 // A CredentialsSource is a source from which provider credentials may be
 // acquired.
-type CredentialsSource string
+type CredentialsSource = common.CredentialsSource
 
 const (
 	// CredentialsSourceNone indicates that a provider does not require
 	// credentials.
-	CredentialsSourceNone CredentialsSource = "None"
+	CredentialsSourceNone = common.CredentialsSourceNone
 
 	// CredentialsSourceSecret indicates that a provider should acquire
 	// credentials from a secret.
-	CredentialsSourceSecret CredentialsSource = "Secret"
+	CredentialsSourceSecret = common.CredentialsSourceSecret
 
 	// CredentialsSourceInjectedIdentity indicates that a provider should use
 	// credentials via its (pod's) identity; i.e. via IRSA for AWS,
 	// Workload Identity for GCP, Pod Identity for Azure, or in-cluster
 	// authentication for the Kubernetes API.
-	CredentialsSourceInjectedIdentity CredentialsSource = "InjectedIdentity"
+	CredentialsSourceInjectedIdentity = common.CredentialsSourceInjectedIdentity
 
 	// CredentialsSourceEnvironment indicates that a provider should acquire
 	// credentials from an environment variable.
-	CredentialsSourceEnvironment CredentialsSource = "Environment"
+	CredentialsSourceEnvironment = common.CredentialsSourceEnvironment
 
 	// CredentialsSourceFilesystem indicates that a provider should acquire
 	// credentials from the filesystem.
-	CredentialsSourceFilesystem CredentialsSource = "Filesystem"
+	CredentialsSourceFilesystem = common.CredentialsSourceFilesystem
 )
 
 // CommonCredentialSelectors provides common selectors for extracting
 // credentials.
-type CommonCredentialSelectors struct {
-	// Fs is a reference to a filesystem location that contains credentials that
-	// must be used to connect to the provider.
-	// +optional
-	Fs *FsSelector `json:"fs,omitempty"`
-
-	// Env is a reference to an environment variable that contains credentials
-	// that must be used to connect to the provider.
-	// +optional
-	Env *EnvSelector `json:"env,omitempty"`
-
-	// A SecretRef is a reference to a secret key that contains the credentials
-	// that must be used to connect to the provider.
-	// +optional
-	SecretRef *SecretKeySelector `json:"secretRef,omitempty"`
-}
+type CommonCredentialSelectors = common.CommonCredentialSelectors
 
 // EnvSelector selects an environment variable.
-type EnvSelector struct {
-	// Name is the name of an environment variable.
-	Name string `json:"name"`
-}
+type EnvSelector = common.EnvSelector
 
 // FsSelector selects a filesystem location.
-type FsSelector struct {
-	// Path is a filesystem path.
-	Path string `json:"path"`
-}
+type FsSelector = common.FsSelector
 
 // A ProviderConfigStatus defines the observed status of a ProviderConfig.
-type ProviderConfigStatus struct {
-	ConditionedStatus `json:",inline"`
-
-	// Users of this provider configuration.
-	Users int64 `json:"users,omitempty"`
-}
+type ProviderConfigStatus = common.ProviderConfigStatus
 
 // A ProviderConfigUsage is a record that a particular managed resource is using
 // a particular provider configuration.
-type ProviderConfigUsage struct {
-	// ProviderConfigReference to the provider config being used.
-	ProviderConfigReference Reference `json:"providerConfigRef"`
-
-	// ResourceReference to the managed resource using the provider config.
-	ResourceReference TypedReference `json:"resourceRef"`
-}
+type ProviderConfigUsage = common.ProviderConfigUsage
 
 // A TargetSpec defines the common fields of objects used for exposing
 // infrastructure to workloads that can be scheduled to.
